+++ packages/fleur/src/Operations.ts
@@ -3,32 +3,34 @@ import {
   OperationContext,
 } from './OperationContext'
 
-export interface OperationDef {
-  (_: OperationContext, ...args: any[]): Promise<void> | void
-}
+export type AnyOperationDef = (
+  _: OperationContext,
+  ...args: any[]
+) => Promise<void> | void
 
-export type OperationType = Operation | OperationDef
+/** Marked as `operations` wrappped Operation type */
+export type Operation = AnyOperationDef
 
-export interface Operation extends OperationDef {
+export type OperationAborter = {
   abort: {
     (context: OperationContext): void
     byKey: (key?: string) => (context: OperationContext) => void
   }
 }
 
-type DefToOperation<T extends OperationDef> = T & Operation
+export type DefToOperation<T extends AnyOperationDef> = T & OperationAborter
 
 export type OperationArgs<T> = T extends (_: any, ...args: infer A) => any
   ? A
   : never
 
 /** Make Operation group from objects */
-export const operations = <T extends { [name: string]: OperationDef }>(
+export const operations = <T extends { [name: string]: AnyOperationDef }>(
   operations: T,
 ): { [K in keyof T]: DefToOperation<T[K]> } => {
   const ops: any = {}
 
-  Object.keys(operations).forEach(key => {
+  Object.keys(operations).forEach((key) => {
     ops[key] = operation(operations[key])
   })
 
@@ -36,8 +38,13 @@ export const operations = <T extends { [name: string]: OperationDef }>(
 }
 
 /** Make one Operation function */
-export const operation = <T extends OperationDef>(op: T): DefToOperation<T> => {
-  const opp = ((context: OperationContext, ...args: any[]) => {
+export const operation = <T extends AnyOperationDef>(
+  op: T,
+): DefToOperation<T> => {
+  const opp: DefToOperation<T> = ((
+    context: OperationContext,
+    ...args: any[]
+  ) => {
     return op(context, ...args)
   }) as DefToOperation<T>
 
@@ -47,10 +54,7 @@ export const operation = <T extends OperationDef>(op: T): DefToOperation<T> => {
 
   abort.byKey = (key?: string) => {
     return (context: OperationContextWithInternalAPI) => {
-      context
-        .getExecuteMap(opp)
-        ?.get(key)
-        ?.abort()
+      context.getExecuteMap(opp)?.get(key)?.abort()
     }
   }