+++ packages/fleur/CHANGELOG.md
@@ -1,5 +1,81 @@
 # @fleur/fleur Changelog
 
+## 3.1.0
+
+### New features
+
+- [#484](https://github.com/fleur-js/fleur/pull/484) Feature: `minOps` introduced!
+- [#484](https://github.com/fleur-js/fleur/pull/484) Feature: Add `context.finally` method
+
+#### minOps
+
+```ts
+import { minOps } from '@fluer/fleur'
+
+export const [SomeStore, someOps] = minOps('SomeDomain', {
+  // Define initial store state.
+  initialState: (): State => ({
+    fetching: false
+    entities: {},
+  }),
+  ops: {
+    async fetchEntity(x, id: string) {
+      // Update store state via `x.commit`.
+      // No more need to define extra actions.
+      // It's shallow merged to state.
+      x.commit({ fetching: true })
+
+      const data = await (await fetch(`/api/entities/${id}`)).json()
+
+      // Or update state with function for deeply set.
+      x.commit((draft) => {
+        draft.fetching = true
+        draft.entities[data.id] = data
+      })
+    },
+    async someAction(x) {
+      // Get the state at the start of the operation can be obtained.
+      x.state
+
+      // Get the latest state with `x.getState()`.
+      x.getState()
+
+      // if you want to mutable type on TypeScript
+      // use `x.unwrapReadonly(state)` (It's no any effect in runtime)
+      //
+      // DON'T MUTATE `state` DIRECTLY WITH `x.unwrapReadonly`!
+      // use `x.commit` if you want!!
+      exioticFunc(x.unwrapReadonly(x))
+    },
+  },
+  /** Can be listened to another domain actions */
+  listens: (on) => [
+    on(anotherDomainActions.action, (draft, payload) => {
+      draft.entities = {}
+    }),
+  ],
+})
+```
+
+#### context.finally
+
+```ts
+minOps('SomeDomain', {
+  initialState: (): State => ({...}),
+  ops: {
+    async fetchData(x, id: string) {
+      const db = await openDB()
+
+      // Look this! `db.close()` called after finish this operation.
+      // It is called even if an exception occurs within the operation.
+      x.finally(() => db.close())
+
+      const data = await db.getItem()
+    },
+  },
+})
+```
+
 ## 3.0.1
 
 - [#496](https://github.com/fleur-js/fleur/pull/496) Annotate this type for typescript-eslint/unbound-method (Fix #447)
@@ -34,7 +110,7 @@ import { Ops } from './ops'
 
 const Component = () => {
   const { executeOperation } = useFleurContext()
-  
+
   executeOperation(Ops.someOp.abort)
   // or `executeOperation(Ops.someOp.abort.byKey(key))` for key specified operation abort